requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
python-dotenv
marshmallow
flask-marshmallow
marshmallow-sqlalchemy
psycopg2-binary
bcrypt